refactor: DI 및 공통 유틸 모듈 도입으로 main.py 코드 정리
- 서비스 및 레포지토리 객체 의존성 주입을 위한 DI 모듈 분리해 main.py 코드 간결화
- main.py에서 공통적으로 쓰이는 성공 응답 반환 함수 정의를 common/response_utils.py로 이동
