+++ Prompting/main.py
@@ -1,53 +1,39 @@
 # MindSync AI Server: FastAPI 기반 회의 지원 서비스
 
 import logging
+from typing import Any
 
 from fastapi import FastAPI, Depends
-from fastapi.responses import JSONResponse
 from fastapi.exceptions import RequestValidationError
 
-from Prompting.schemas import RoomIdRequest, ChatRequest, AgendaRequest, Response, ChatResponse
+from Prompting.schemas import RoomIdRequest, ChatRequest, AgendaRequest, Response
 from Prompting.repository import AgendaRepository, ChatRepository, RoomRepository, UserRepository
 from Prompting.services import AgendaGenerator, MeetingSummarizer, MbtiChatGenerator
-from Prompting.services.context_builders.meeting_history_builder import MeetingHistoryBuilder
-from Prompting.usecases.summarize_usecase import load_summary_context
-from Prompting.usecases.mbti_chat_usecase import load_chat_context
+from Prompting.usecases import load_summary_context, load_chat_context
 
 from Prompting.exceptions.errors import GeminiCallError, GeminiParseError, MongoAccessError, PromptBuildError
+from Prompting.exceptions.decorators import catch_and_raise
 from Prompting.exceptions.handlers import custom_exception_handler, request_validation_exception_handler, general_exception_handler
 
+from Prompting.common.resonse_util import success_response
 
-# -------------------- 기본 설정 ------------------------
+from .di import (
+    get_agenda_repo, get_room_repo, get_chat_repo, get_user_repo,
+    get_agenda_service, get_bot_service, get_summarizer_service
+)
+
+
+# 기본 설정 및 예외 핸들러 등록 ------------------------------------------------------------------------
 logging.basicConfig(level=logging.INFO)  # 로깅 설정
 app = FastAPI()  # FastAPI 애플리케이션 생성
 
-# -------------------- 전역 핸들러 등록 --------------------
 for exc in [GeminiCallError, GeminiParseError, MongoAccessError, PromptBuildError]:
     app.add_exception_handler(exc, custom_exception_handler)
 app.add_exception_handler(RequestValidationError, request_validation_exception_handler)
 app.add_exception_handler(Exception, general_exception_handler)
 
-# -------------------- 공통 유틸 함수 --------------------
-def success_response(data=None, message="요청이 성공했습니다."):
-    return JSONResponse(status_code=200, content={"status": "SUCCESS", "message": message, "data": data})
 
-# -------------------- DI 함수 -------------------------
-def get_agenda_service():
-    return AgendaGenerator()
-def get_summarizer_service():
-    return MeetingSummarizer()
-def get_bot_service():
-    return MbtiChatGenerator()
-def get_agenda_repo():
-    return AgendaRepository()
-def get_chat_repo():
-    return ChatRepository()
-def get_room_repo():
-    return RoomRepository()
-def get_user_repo():
-    return UserRepository()
-
-# -------------------- 엔드포인트 --------------------
+# API 엔드포인트 -------------------------------------------------------------------------------------
 @app.post("/agenda_generation/", response_model=Response)
 async def generate_and_save_agendas(
         request: AgendaRequest,
@@ -100,7 +86,7 @@ async def summarize_meeting_chat(
     summary_data = summarizer.parse_response_to_summary_data(summary)
     await room_repo.save_summary(room_id=request.roomId, summary=summary_data)
 
-    return success_response(data=summary, message="요약 생성을 완료했습니다.")
+    return success_response(data=summary_data, message="요약 생성을 완료했습니다.")
 
 
 @app.post("/mbti_chat/", response_model=Response)
@@ -134,9 +120,5 @@ async def generate_mbti_chat(
 
 @app.get("/")
 def root():
-    """
-    루트 경로 핸들러
-
-    :return: 환영 메시지, dict
-    """
+    """루트 경로 핸들러"""
     return {"message": "Hello It's MindSync AI service server"}
